refactor(server): moved paths to new config
