+++ server/src/config/IPaths.ts
