+++ server/src/config/defaultPaths.ts
@@ -1,7 +1,14 @@
 import path from 'path';
-import IServerConfig from './IServerConfig';
+import IPaths from './IPaths';
 
-export default (userData: string, tempData: string): IServerConfig => {
+/**
+ * Creates the paths required for H5P to work.
+ * @param userData a directory which is scoped per user and persistent
+ * @param tempDir a temporary directory whose content can be deleted after the
+ * app was closed
+ * @returns
+ */
+export default (userData: string, tempData: string): IPaths => {
     return {
         librariesPath: path.join(userData, 'libraries'),
         cache: path.join(userData, 'store.json'),
